Updated the Grid component styles
and moved util function to a separate file.
